+++ test/test_util.py
@@ -13,12 +13,11 @@ from urllib.parse import urlparse
 
 import pytest
 
-from urllib3 import add_stderr_logger, disable_warnings, util
+from urllib3 import add_stderr_logger, disable_warnings
 from urllib3.connection import ProxyConfig
 from urllib3.exceptions import (
     InsecureRequestWarning,
     LocationParseError,
-    SNIMissingWarning,
     TimeoutStateError,
     UnrewindableBodyError,
 )
@@ -799,7 +798,12 @@ class TestUtil:
 
     def test_connection_requires_http_tunnel_http_proxy(self) -> None:
         proxy = parse_url("http://proxy:8080")
-        proxy_config = ProxyConfig(ssl_context=None, use_forwarding_for_https=False)
+        proxy_config = ProxyConfig(
+            ssl_context=None,
+            use_forwarding_for_https=False,
+            assert_hostname=None,
+            assert_fingerprint=None,
+        )
         destination_scheme = "http"
         assert not connection_requires_http_tunnel(
             proxy, proxy_config, destination_scheme
@@ -810,7 +814,12 @@ class TestUtil:
 
     def test_connection_requires_http_tunnel_https_proxy(self) -> None:
         proxy = parse_url("https://proxy:8443")
-        proxy_config = ProxyConfig(ssl_context=None, use_forwarding_for_https=False)
+        proxy_config = ProxyConfig(
+            ssl_context=None,
+            use_forwarding_for_https=False,
+            assert_hostname=None,
+            assert_fingerprint=None,
+        )
         destination_scheme = "http"
         assert not connection_requires_http_tunnel(
             proxy, proxy_config, destination_scheme
@@ -1027,21 +1036,6 @@ class TestUtilSSL:
                 server_hostname=server_hostname,
             )
         return mock_context, warn
-
-    def test_ssl_wrap_socket_sni_hostname_use_or_warn(self) -> None:
-        """Test that either an SNI hostname is used or a warning is made."""
-        sock = Mock()
-        context, warn = self._wrap_socket_and_mock_warn(sock, "www.google.com")
-        if util.HAS_SNI:
-            warn.assert_not_called()
-            context.wrap_socket.assert_called_once_with(
-                sock, server_hostname="www.google.com"
-            )
-        else:
-            assert warn.call_count >= 1
-            warnings = [call[0][1] for call in warn.call_args_list]
-            assert SNIMissingWarning in warnings
-            context.wrap_socket.assert_called_once_with(sock)
 
     def test_ssl_wrap_socket_sni_ip_address_no_warn(self) -> None:
         """Test that a warning is not made if server_hostname is an IP address."""