+++ test/with_dummyserver/test_socketlevel.py
@@ -8,6 +8,7 @@ import select
 import shutil
 import socket
 import ssl
+import sys
 import tempfile
 import time
 from collections import OrderedDict
@@ -84,8 +85,6 @@ class TestCookies(SocketDummyServerTestCase):
 
 class TestSNI(SocketDummyServerTestCase):
     def test_hostname_in_first_request_packet(self) -> None:
-        if not util.HAS_SNI:
-            pytest.skip("SNI-support not available")
         done_receiving = Event()
         self.buf = b""
 
@@ -1492,6 +1491,63 @@ class TestSSL(SocketDummyServerTestCase):
                 pool.request("GET", "/", retries=False, timeout=LONG_TIMEOUT)
         assert server_closed.wait(LONG_TIMEOUT), "The socket was not terminated"
 
+    # SecureTransport can read only small pieces of data at the moment.
+    # https://github.com/urllib3/urllib3/pull/2674
+    @notSecureTransport()
+    @pytest.mark.skipif(
+        os.environ.get("CI") == "true" and sys.implementation.name == "pypy",
+        reason="too slow to run in CI",
+    )
+    @pytest.mark.parametrize(
+        "preload_content,read_amt", [(True, None), (False, None), (False, 2**31)]
+    )
+    def test_requesting_large_resources_via_ssl(
+        self, preload_content: bool, read_amt: Optional[int]
+    ) -> None:
+        """
+        Ensure that it is possible to read 2 GiB or more via an SSL
+        socket.
+        https://github.com/urllib3/urllib3/issues/2513
+        """
+        content_length = 2**31  # (`int` max value in C) + 1.
+        ssl_ready = Event()
+
+        def socket_handler(listener: socket.socket) -> None:
+            sock = listener.accept()[0]
+            ssl_sock = ssl.wrap_socket(
+                sock,
+                server_side=True,
+                keyfile=DEFAULT_CERTS["keyfile"],
+                certfile=DEFAULT_CERTS["certfile"],
+                ca_certs=DEFAULT_CA,
+            )
+            ssl_ready.set()
+
+            while not ssl_sock.recv(65536).endswith(b"\r\n\r\n"):
+                continue
+
+            ssl_sock.send(
+                b"HTTP/1.1 200 OK\r\n"
+                b"Content-Type: text/plain\r\n"
+                b"Content-Length: %d\r\n\r\n" % content_length
+            )
+
+            chunks = 2
+            for i in range(chunks):
+                ssl_sock.sendall(bytes(content_length // chunks))
+
+            ssl_sock.close()
+            sock.close()
+
+        self._start_server(socket_handler)
+        ssl_ready.wait(5)
+        with HTTPSConnectionPool(
+            self.host, self.port, ca_certs=DEFAULT_CA, retries=False
+        ) as pool:
+            response = pool.request("GET", "/", preload_content=preload_content)
+            data = response.data if preload_content else response.read(read_amt)
+            assert len(data) == content_length
+
 
 class TestErrorWrapping(SocketDummyServerTestCase):
     def test_bad_statusline(self) -> None:
@@ -1643,6 +1699,75 @@ class TestHeaders(SocketDummyServerTestCase):
                 (k, v) for (k, v) in r.headers.items() if k.startswith("X-Header-")
             ]
             assert expected_response_headers == actual_response_headers
+
+    @pytest.mark.parametrize(
+        "method_type, body_type",
+        [
+            ("GET", None),
+            ("POST", None),
+            ("POST", "bytes"),
+            ("POST", "bytes-io"),
+        ],
+    )
+    def test_headers_sent_with_add(
+        self, method_type: str, body_type: Optional[str]
+    ) -> None:
+        """
+        Confirm that when adding headers with combine=True that we simply append to the
+        most recent value, rather than create a new header line.
+        """
+        body: Union[None, bytes, io.BytesIO]
+        if body_type is None:
+            body = None
+        elif body_type == "bytes":
+            body = b"my-body"
+        elif body_type == "bytes-io":
+            body = io.BytesIO(b"bytes-io-body")
+            body.seek(0, 0)
+        else:
+            raise ValueError("Unknonw body type")
+
+        buffer: bytes = b""
+
+        def socket_handler(listener: socket.socket) -> None:
+            nonlocal buffer
+            sock = listener.accept()[0]
+            sock.settimeout(0)
+
+            start = time.time()
+            while time.time() - start < (LONG_TIMEOUT / 2):
+                try:
+                    buffer += sock.recv(65536)
+                except OSError:
+                    continue
+
+            sock.sendall(
+                b"HTTP/1.1 200 OK\r\n"
+                b"Server: example.com\r\n"
+                b"Content-Length: 0\r\n\r\n"
+            )
+            sock.close()
+
+        self._start_server(socket_handler)
+
+        headers = HTTPHeaderDict()
+        headers.add("A", "1")
+        headers.add("C", "3")
+        headers.add("B", "2")
+        headers.add("B", "3")
+        headers.add("A", "4", combine=False)
+        headers.add("C", "5", combine=True)
+        headers.add("C", "6")
+
+        with HTTPConnectionPool(self.host, self.port, retries=False) as pool:
+            r = pool.request(
+                method_type,
+                "/",
+                body=body,
+                headers=headers,
+            )
+            assert r.status == 200
+            assert b"A: 1\r\nA: 4\r\nC: 3, 5\r\nC: 6\r\nB: 2\r\nB: 3" in buffer
 
 
 class TestBrokenHeaders(SocketDummyServerTestCase):